+++ NEWS.md
@@ -1,3 +1,8 @@
+# methylGSA 0.99.14
+* Included gene sets description
+* Included visualization
+* Included option for analyzing promoter/body only
+
 # methylGSA 0.99.10 
 * Support user-supplied mapping between CpGs and Genes